refactor: remove last logicalPlan implementation
